pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-supabase-power"
version = "1.0.0"
description = "Powerful ComfyUI nodes for Supabase integration - Database, Storage, Realtime, Auth, Edge Functions"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ComfyUI-Supabase-Power Contributors"}
]
keywords = ["comfyui", "supabase", "database", "storage", "realtime", "nodes"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "supabase>=2.0.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["comfyui_supabase_power*"]

[tool.setuptools.package-data]
"*" = ["*.js", "*.css"]
